requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
boto3==1.24.36
botocore==1.27.36
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.51.0
django-cors-headers==3.8.0
django-environ==0.8.1
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-smtp-ssl==1.0
django-storages==1.12.3
djangorestframework==3.12.4
filelock==3.7.0
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
numpy==1.23.0
oauthlib==3.2.1
Pillow==8.3.2
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.7.0
pycparser==2.21
PyJWT==2.5.0
python-dateutil==2.8.2
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2021.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
types-cryptography==3.3.23
tzdata==2021.5
urllib3==1.26.10
virtualenv==20.14.1

